// types.ts
export type Observation = {
  timestamp_hkt: string;
  date: string;
  day: string;
  platform: string;
  stock: string;
  token_symbol: string;
  friday_close: string;
  token_price: string;
  bid: string;
  ask: string;
  spread_pct: string;
  volume_24h: string;
  high_24h: string;
  low_24h: string;
  vs_friday_pct: string;
  run_id: string;
  data_status: string;
};